Music_Generation/requirements.txt
PyGithub
python-box
ensure
joblib
numpy
scikit-learn
tqdm
librosa
keras
tensorflow
Flask